feat: add build info to version
This is the first implementation of adding build information to the
--version argument of the command line tool. Example output when
`e4e-data-management-rust --version` is run looks like:
```
version: <version number>
git commit hash: <commit hash>
build time: <build time>
```
The above string is currently stored in its entirety in a text file
inside the build directory, which is necessary since clap requires
strings to be static to generate the output. Ideally this information
should be split up to make things a bit cleaner.