ci: trigger workflow on all pushes and pull requests for any branch
The commit hash appended to the generated binaries should be able to map to a
commit that exists on the commit tree. This allows binaries for work-in-progress
features to also be generated and tested.

Note that the hash from a pull request will not exist on the commit tree since
the merge technically does not exist yet.