Adding git & build info to --version (#15)
* feat: add build info to version

This is the first implementation of adding build information to the
--version argument of the command line tool. Example output when
`e4e-data-management-rust --version` is run looks like:
```
version: <version number>
git commit hash: <commit hash>
build time: <build time>
```
The above string is currently stored in its entirety in a text file
inside the build directory, which is necessary since clap requires
strings to be static to generate the output. Ideally this information
should be split up to make things a bit cleaner.

* chore: removed redundant code in build.rs

* fix: removed write to file, now storing version info in environment variables

* fix: using ? operator in build.rs instead of manually unwrapping

* fix: run clippy

* style: run cargo fmt

* ci: add build for multiple oses, upload build artifacts

* ci: remove macos m1 instance

* ci: change checkout and upload-artifact actions from v3 to v4

* ci: add current commit hash to binary names

* fix: get short hash without requiring git to support windows-latest

* ci: trigger workflow on all pushes and pull requests for any branch

The commit hash appended to the generated binaries should be able to map to a
commit that exists on the commit tree. This allows binaries for work-in-progress
features to also be generated and tested.

Note that the hash from a pull request will not exist on the commit tree since
the merge technically does not exist yet.